Control.Exception.Throwable

Description

The mostly exceptions has the field for its cause, and its clue.

The cause is the message.

The clue is some stuff for find bugs, the clue can be omitted if you don't need it. (e.g. ioException', illegalArgumentException)

Documentation

data GeneralException a

An other of these exceptions

Constructors

GeneralException

Fields

exceptionName :: String

This should be named by PascalCase for show, for example: MyOperation, Database

exceptionCause :: String

The message for the cause of the exception

exceptionClue :: a

The clue of the exception. This can be anything of Show instance

generalException :: String -> String -> GeneralException ()

A constructor for GeneralException but doesn't take the clue

data IOException' a

Constructors

IOException'

Fields

iOException'Cause :: String

iOException'Clue :: a

FileNotFoundException

Fields

fileNotFoundExceptionCause :: String

fileNotFoundExceptionClue :: a

ioException' :: String -> IOException' ()

data IllegalArgumentException a

Constructors

IllegalArgumentException

Fields

illegalArgumentExceptionCause :: String

illegalArgumentExceptionClue :: a

illegalArgumentException :: String -> IllegalArgumentException ()

data IndexOutOfBoundsException a

Constructors

IndexOutOfBoundsException

Fields

indexOutOfBoundsExceptionCause :: String

indexOutOfBoundsExceptionClue :: a

indexOutOfBoundsException :: String -> IndexOutOfBoundsException ()