Control.Exception.Throwable.TH

Description

The data types creator of exceptions in the compile type.

It has the record of "cause" and "clue".

Documentation

declareException :: DatatypeName -> [ValueConstructorName] -> DecsQ

Declare simple concrete exception data type in the compile time.

If the empty list is given to constrNames, create empty data type.

typeName and constrNames must be PascalCase (e.g> IOException', IndexOutOfBoundsException. NG> "ioException'", "indexOutOfBoundsException") .